Change battle score type

+++ db/migrate/20190326201124_add_winner_loser_and_scores_to_battles.rb
@@ -4,7 +4,7 @@ class AddWinnerLoserAndScoresToBattles < ActiveRecord::Migration[5.2] #:nodoc:
   def change
     add_column :battles, :winner, :integer
     add_column :battles, :loser, :integer
-    add_column :battles, :winner_score, :integer
-    add_column :battles, :loser_score, :integer
+    add_column :battles, :winner_score, :float
+    add_column :battles, :loser_score, :float
   end
 end
